ai_notes_translation/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
googletrans==4.0.0rc1
psycopg2-binary>=2.9.0
python-decouple>=3.8
django-redis>=5.2.0
redis>=4.5.0

